+++ condow_core/CHANGELOG.md
@@ -4,7 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.11.0] -  - 2021/12/11
+## [0.12.0] - unreleased
+
+### CHANGED
+
+- [BREAKING] Config does no more implement equality traits (there is now an f64 in there)
+- [BREAKING] Added a new field for retry config to the Config struct
+- [BREAKING] Config is now non-exhaustive
+
+## [0.11.0] -  2021/12/11
 
 ### Fixed